[REF] mail: move firstUnreadMessage to channel model
